Handle unaligned areas

+++ popcnt/subs.h
@@ -1,2 +1,4 @@
-extern unsigned int cpuidfn(unsigned int);
-extern unsigned int countbits(uint16_t *a, size_t n);
+extern unsigned int cpuidfn(unsigned long int);
+
+/* call with 64-bit aligned address and n evenly divisible by 8 */
+extern unsigned int countbits(uint16_t *a, unsigned long n);
